added basic set up scripts, and mountaineer avatar
Character Scripts
Character Animation => a pretty simple animation script to avoid using the unity animator which is clunky af
Character State => another simple script used only to "store" values about the character (e.g. energy, class name, level etc)

HUD Scripts
A bunch of scripts to attach to HUD objects, most are empty but some important ones are

HUDPortrait => the script for the portrait of the selected mountaineer
HUDGameOver => displays when the losing the game

add and delete as many of these as you need, but the idea is that each component in the HUD feeds its information into the parent HUD script so that all the information is accessible in one place.

World Scripts:
GameRules Script => for rules like "number of actions per turn", "game duration" etc (like character state, but stores values about the game)

Background Music => simple script that controls the background music with varying levels of intensity

World Renderer => currently empty, but is useful for things like procedural generation and stuff if we chose to go that route